requirements.txt
Django==2.2.28
django-webpack-loader==0.6.0
djangorestframework==3.13.1
django_cors_headers==3.13.0
pytz==2022.1
coverage==6.5
selenium==4.4.0
webdriver_manager==3.8.4
packaging==21.3